WIP: 12: part 1 - algorithm no worky
Implement the transforms. But, the scoring algorithm is no good, you
quickly end up in a dead-end with no best next step and it fails.

So, probably, it needs to be a recursive situation; but I have no idea
how that's going to be efficient - I guess it needs to be guided by a
score